Refactor memory.rs into modular structure
- Split large memory.rs file (1604 lines) into logical modules:
  - memory/mod.rs: Main module with re-exports
  - memory/ptr.rs: Pointer types (CallerAllocatedPtr, CalleeAllocatedPtr)
  - memory/wstring.rs: Wide string types (CallerAllocatedWString, CalleeAllocatedWString)
  - memory/array.rs: Array types (CallerAllocatedArray, CalleeAllocatedArray)
  - memory/ptr_array.rs: Pointer array types (CallerAllocatedPtrArray, CalleeAllocatedPtrArray)
  - memory/tests.rs: All test functions

- Maintained all existing functionality and API
- All tests pass (29 tests)
- All examples work correctly
- No clippy warnings
- Improved code organization and maintainability
- Each module focuses on a specific type category
- Easier to navigate and understand the codebase
